+++ CHANGELOG.md
@@ -12,7 +12,7 @@
 
 ## 1.4.0
 
-- **Major Update** Extracted the get_connect module from the package and published it as refreshed_connect.
+- **Major Update** Extracted the get_connect module from the package and published it as vaultify.
 - Removed infrequently used string extension methods
 
 ## 1.3.0